PP-4637: Use EncryptedPaymentData as a parameter in the GooglePayDecrypter
@oswaldquek
