--- src/test/resources/googlepay/google-pay-response.json
@@ -1,14 +1,0 @@
-{
-  "tokenizationData": {
-    "type": "DIRECT",
-    "token": {
-      "signature":"..."
-    },
-    "intermediateSigningKey": {
-      "signedKey":"{\"keyValue\":\"...\",\"keyExpiration\":\"...\"}",
-      "signatures":["..."]
-    },
-    "protocolVersion":"ECv2",
-    "signedMessage":"{\"encryptedMessage\":\"...\",\"ephemeralPublicKey\":\"...\",\"tag\":\"...\"}"
-  }
-}
